Hopefully fix the volume problem for non-root user
Since I want to use USER perforce, I need to make sure I can actually
write to the P4ROOT directory. Docker on Mac mounts volume as user 1000
while perforce in this docker instance has id 103

So we hack passwd to change the id for perforce to 1000, and if by magic
the volume is own by user perforce.